Add `cljrs ir-prebuild` subcommand for pre-lowering namespaces to IR
Splits cljrs-ir-prebuild into a lib (run_prebuild) + thin binary so the
existing standalone tool and the new `cljrs ir-prebuild build`/`dump`
subcommands share one implementation instead of duplicating it. Bundles
are loaded back at startup via the existing public
`cljrs_eval::load_prebuilt_ir` API, letting pre-lowered functions skip
warmup and execute at Tier 1 from their first call — useful for
cold-start-sensitive targets like a wasm embedder. Adds the missing
crate README for cljrs-ir-prebuild and documents the new subcommand.
